Split main CLI tests into a sibling module
Move the binary-only main_cli test module into src/main_cli_tests.rs so
src/main_cli.rs stays below the repository's 400-line file limit.

While moving the tests, replace the duplicated Worker, ClaudeBridge,
and Onboard passthrough cases with a single rstest parameterized case.
This keeps the same behavioural coverage while reducing local test
repetition.
